refactor(error-handling): Escalate ExitInfo returned by PlatformInconsistencyCheckerWorker::checkIfPathAndNameAreValid instead of swallowing it (Clément's review comments)

+++ test/libcommon/utility/testutility.cpp
@@ -476,7 +476,7 @@ void TestUtility::testGenerateRandomNumber() {
     bool allValuesAreTheSame = true;
     for (auto i = 0; i < 100; i++) {
         const auto val = CommonUtility::generateRandomNumber(1, 100);
-        if (val != sequence[i]) {
+        if (val != sequence[static_cast<size_t>(i)]) {
             allValuesAreTheSame = false;
             break;
         }
